third_party/blink/web_tests/platform/generic/html/document_metadata/head-check-expected.txt

This test checks whether one (and only one) HEAD element is created for various HTML documents. See Bug 7321 REGRESSION: style tag in body causes two head elements to appear in the DOM.

PASS successfullyParsed is true

TEST COMPLETE

Testing: head-check-1.html
PASS iframe.contentWindow.document.getElementsByTagName('head').length is 1
PASS iframe.contentWindow.document.firstChild.firstChild.nodeName.toLowerCase() is 'head'


Testing: head-check-2.html
PASS iframe.contentWindow.document.getElementsByTagName('head').length is 1
PASS iframe.contentWindow.document.firstChild.firstChild.nodeName.toLowerCase() is 'head'


Testing: head-check-3.html
PASS iframe.contentWindow.document.getElementsByTagName('head').length is 1
PASS iframe.contentWindow.document.firstChild.firstChild.nodeName.toLowerCase() is 'head'


Testing: head-check-4.html
PASS iframe.contentWindow.document.getElementsByTagName('head').length is 1
PASS iframe.contentWindow.document.firstChild.firstChild.nodeName.toLowerCase() is 'head'


TEST COMPLETE
